pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-analyzer"
version = "1.0.0"
description = "Multi-repo code graph builder and cross-repo call analyzer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Rikiz"}
]
keywords = ["neo4j", "code-analysis", "graph", "multi-repo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Database :: Database Engines/Servers",
]
dependencies = [
    "neo4j>=5.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
repo-analyzer = "repo_analyzer.main:main"

[project.urls]
Homepage = "https://github.com/Rikiz/graph_extractor_by_opencode"
Repository = "https://github.com/Rikiz/graph_extractor_by_opencode"
Documentation = "https://github.com/Rikiz/graph_extractor_by_opencode#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["repo_analyzer*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
